Added a branch for developing custom synchronization for the hpx runtime.
This entails:
1) A set of low-level synchronization routines.
2) An implementation of a bunch of different shared-memory synchronization objects.

(1) is done using a platform-specific sync.h header in the sync directory. This defaults to using gcc's __atomic builtins (and is modeled based on their interface), which is defined in gcc_builtin_atomic.h. Other options are to use the __sync builtin interface, and to use a craycc __sync interface, and to use a library interface. Only the __atomic one is implemented right now.

(2) is started using the locks.h header, which only implements a simple TATAS lock (actually a backoff lock). Fastpaths are inlined, but slowpaths are implemented in locks.c.

No tests are yet included.