Primary change was to move from string keys to hashed keys in the parcel structure.
This isn't perfect, for example, the MD5 keys might collide, but it's better for everything else. This had a serious cascade effect which I dealt with in one major way, I extracted a bunch of things into their own files. Primarily in the src/parcel directory. If this becomes a performance problem, then we can deal with it later.