Fixed compilation for parcelhandler.
A quick look at this file makes it appear to have some questionable things going on, so it wouldn't surprise me if it didn't really work. I certainly broke some tests with the changes that I made.

On specific concern is action registration. My understanding is that 'check' runs each test using fork, in which case it is ok to have action registration in each test. If this is not using fork, then it's not going to work.