Swtich libhpx source to correctly deal with config.h, rather than relying on Makefile.am to -D things we're already AC_DEFINING.
Also pushed the network.h header into the src directory, it shouldn't be a public header. This wasn't a big deal, but required me to touch a bunch of files to update the include path. I also hadto #include <config.h> in the places in src/* that use HAVE_*.

A side effect of this is that the public headers don't have the HAVE_* tests, so the example Makefile.ams don't need the HAVE_* logic anymore. Everything is simpler now.