Gave the responsibility of sending parcels to the parcelhandler, and added an interface for it.
The current parcelhandler_send interface doesn't do anything, we'll need to
fix that. In addition, the current parcel-send interface does not correctly deal
with local parcels because it can't yet disambiguate local virtual
addresses. This needs to be updated to check the destination locality for the
meantime, until we have a global virtual address space.