Removed thread_from_sp().
Each worker now keeps track of the current thread as a thread-local variable. When we transfer, the transfer continuation needs to exchange the previous current thread for the thread we're transferring to. Because we're cooperative, this works fine. This means that we no longer care if all of the thread stacks are the same size, or even contiguous, because we never have to to math on the stack address.