Updated startup and shutdown paths.
Startup now makes the main() thread a worker in the scheduler, rather than putting it to sleep. We also allow workers to self-assign ids, based on an atomic integer.

Shutdown now makes the network shutdown in addition to the scheduler. When a network gets a shutdown request, it performs a network-network communication to broadcast the shutdown to the entire system. On receive, the network checks to see if the message is a parcel, or a code, and executes the associated functionality directly if there is a code. For now, the only code available is NETWORK_SHUTDOWN.

The network shutdown path causes MPI to complain at the moment, so I still have to figure out some sort of timing problem.