Simplified shutdown further.
Now, we just shut down the scheduler, and the prime thread can then shutdown the network. If the shutdown request comes over the network, then we set the network to shutdown before we shutdown the scheduler, which prevents us from re-broadcasting the scheduler.

There is still the MPI complaint at shutdown about ordering, which needs to be debugged.