Normalized the counter implementation.
The counter implementation was doing a bunch of things that weren't really necessary. I stripped it down to the bare bones. In particular, the counter has an LCO, thus doesn't need a future. I also switched it to be a countdown counter, rather than counting up to a limit, which reduces its size by 8 bytes. It can be expanded a bit if the waiting threads really need to know the original limit. I didn't test the new implementation, but it should be quite close to working.