Formalized the network event and state model as enums. Modified shutdown so that the network_progress() loop is actually the one to broadcast shutdown.
This ensures that the _network_cancel() operation happens before the shutdown broadcast, and that a FUNNELLED network will work correctly. The mechanism to do this is to add an additional state to the network which is that a shutdown is pending, but not yet performed. Then, in progress, the broadcast only occurs if the progress() loop can switch the state from pending to shutdown.