Moved the hashtable into libsync. Preparing for a real concurrent hashtable.
The PGAS map has to be a concurrent hashtable, and it has to be relatively quick to search. I will implement a concurrent hashtable based on the optimistic cuckoo hash presented by Fan, Andersen, and Kaminsky---http://dl.acm.org/citation.cfm?id=2482662. Right now, the hash is just a UTHASH protected by a single global lock. This is terrible for scalability, but allows the benchmarks to work.