Added an interface to index into a global array of futures.
This is necessary, because the future_t is not exposed in the hpx.h header, so the application programmer can't do the math requires for hpx_addr_add. We could eliminate it (in fact, we could likely eliminate special purpose future array allocation) by exposing the size of a future, in bytes, in the hpx.h header.