Fixes #1. (segv in debug mode)
When --enable-debug flag is on, hpx_get_rank() segvs before
boot because boot_rank is called on a NULL pointer. We could
have had a different log function for pre-boot messages that
doesn't print the rank, but for now I added a check in
hpx_get_rank() to avoid the seg fault.