Discovered another bug in the future, where a lock was not being released correctly under certain circumstances.
After this update, FMM runs to completion. I don't understand why this would have been causing the problem I was seeing, but I'll take it for now. I have a suspicion that there's something else that might have been causing the problem though, that may still be in there, and the incorrect synchronization was merely exposing it.