Added a stack pointer to the parcel structure as it seems like the most efficient way to support lazy stack binding.
All threads already have parcels, so why not use the parcel structure as the "thread control block" as well. The primary concern here is that parcels are block allocated on cache-line boundaries, so adding fields to them will decrease the minimum payload size that we can allocate. Note that this doesn't have anything to do with the amount of data we send on the network.