Added an hpx_thread_get_tls_id() to the public interface so that lightweight threads can get integers assigned to them.
The hpx_get_my_thread_id() is not a lightweight thread call, it is a topology call. It returns a dense integer related to the current scheduler thread, not the current user thread. Some applications that have long-running threads occasionally want to number lightweight threads. When a lightweight thread calls this function for the first time, it will be assigned a unique, dense integer within the context of the current scheduler structure. All future calls within the same thread will be given the same ID. There is no way to recover IDs, and there are only 2^31 ids at this point, so they should be used carefully.