Complete rewrite of the hpx.c file, and cascading changes everywhere else.
I exposed the locality as a first-class object, in a header. This allows direct access to the various parts of the locality without function calls or clutter resulting from threading things around the system. Unfortunately, this caused a large number of changes everyhwere. This checkin compiles, but is untested for behavior. I'm 99% sure that it will segfault.