Fix MPI for the rethread branch.
The problem was with hpx_shutdown(). The semantics of hpx_bcast have changed--the
calling locality also broadcasts the parcel to itself. The way it was set up was
that the root locality broadcasted a parcel to all localities (but itself), flushed
the transport and then shut itself down. I have moved the transport flush in the
progress thread now, so that it sends out the pending parcels before shutting itself
down.