Add a hpx_move operation to hpx.h
This simply calls the btt_remap() operation in the btt class. hpx_move()
could be a high-level operation like hpx_global_alloc, because it would
likely be similar for AGAS-software and AGAS-switch, but currently we
need it to be a no-op for PGAS. The easiest way to do this is to add a
remap operation to the BTT class.