Use the hpx_thread_continue_cleanup() interface to free a block once it's been invalidated.
In order for this to work, we have to allocate blocks individually in the locality's global block allocation routine. We used to allocate blocks as one big contiguous allocation, which was nice for performance reasons. Unfortunately, this option doesn't work with AGAS. This should be a decision made based on the GAS implementation, but I'm not worried about it for now.