Updated the network to use the two-lock queue.
When we didn't have any local workstealing, it was important that the work-sharing queue deal well with high contention, so we were using an M&S queue. Now that we have work stealing, the queue is not a point of contention. Switching the default queue to an M&S two-lock queue doesn't add any overhead for infrequent access to the network (i.e., fibonacci), but cuts pingpong latencies dramatically, and cuts distributed fmm execution times in half (for 100000x100000 on cutter with 2 nodes).