Do a rank-based array initialization for futures.
We could do a better job if we had a version of hpx_global_alloc that takes the initializer, because we know that the blocks can't move around during allocation. As it is, we still need to user parcels and mapping to initialize all of the blocks, because they might have moved. Still better than it was though.