Updated sendrecv to more accurately reflect the overlap of compuation and communication.
Since HPX only exposes the send async function, it makes the most sense to measure overlap in the sender thread, rather than in the receiver thread as is done in MPI. I adjusted the sendrecv test to do "Isend, worker, wait" rather than an attempted Irecv version. We could also put the worker() loop in the receiver, but I'm not sure that that is the same test.