Removed stack caching for now.
Stack caching, as it was implemented, was causing pathological behavior in some cases. This patch removes stack caching---it is necessary to use a good malloc to get good performance where thread creation and termination times are on the critical path, like fibonacci. Quick testing shows that jemalloc is roughly equivalent to stack caching, avoids the pathological behavior seen previously, and is trivial to LD_PRELOAD on the systems we currently use.