Fixed a bug in lulesh's initialization.
The domain's symmX, symmY, and symmZ fields were sometimes left uninitialized, which lead to erroneous behavior in lulesh-comp.c, where they are tested against 0. The fix to this problem was to initialize these fields to NULL if necessary.

It's not entirely clear to me why this only created problems in some execution orderings, but it was clearly wrong and I haven't seen any issues with the updated version.