Fixed the allgather example.
The problem with the allgather turns out to be a mismatch between the semantics of setting a local allgather (synchronous-with-wait-for-phase) vs. setting a remote allgather (asynchronous-with-proxy-wait). The allgather doesn't cycle as one would expect if you issue a remote setid via the proxy, and then issue the get operation. I'm not sure why this doesn't work, we'll have to rationalize this further, however waiting for rsync for the setid fixes our issues for now.