Fixed hpx_gas_alloc.
hpx_gas_alloc() only allows a single block to be allocated at a time. The runtime maintains a
private sbrk pointer for these allocations that starts from the top of the address space and
is bumped down for each allocation. gas_alloc'ed memory has affinity to the allocating
locality, by default. If the local address space runs out of space, these requests should
be forwarded to another (random) locality.