Completed two pending changes to process termination.
1. Untracked processes now work correctly. We need to be able to tell
whether a process is tracked or not from its PID. Since processes
identifiers are derived out of a processes's global address, all untracked
"processes" are currently assigned PID = 0. All RPC calls operating on
a process check for this condition to determine if a process is a real
(tracked) process or not.

2. Fixed multiple attachs (hpx_process_call) on a single process.