Added an option for maximum number of send/receive transport requests.
The same limit is used for both send and receive requests. Right now,
the only 2 transports that make use of this config option are MPI and
Photon. SSSP example has been modified to accept this option and it
has been tested on cutter with MPI (on qLogic) and Photon (on
Mellanox).

The number of maximum requests seems to have impact on performance,
based on preliminary testing. More importantly, PSM limits the
maximum number of sends and receives to 1048576.

Signed-off-by: Marcin Zalewski <zalewski@indiana.edu>