Force the call to the continuation in thread_transfer() to occur with a 16-byte aligned stack pointer.
It occasionally turns out that thread_transfer() is called with an 8-byte
aligned stack pointer. It's not clear why this would happen, as it is our
understanding that gcc is responsible for correctly aligning $rsp before the
call, but we clearly see this not happening, e.g., in the asm for
_light_network_action, gcc subtracts 8 bytes from rsp and then calls
thread_yield(). We don't know why this occurs, but we know that it doesn't
matter unless the continuation uses 16-byte aligned operations to the stack.

Unfortunately, gcc does generate movaps operations to the stack in some
optimization situations, and we see segfaults. This re-alignment that has been
added should suppress that bug.