Add the embedded photon and jemalloc symbols to libhpx.a.
I changed the embedded jemalloc and photon .la targets to be noinst, which makes
them "convenience" libraries. Libtool will archive link them together with
everything else in libhpx.a, so that users can just -lhpx and get all of the
symbols that they need.

verbs-users still need the verbs libraries since they are linked dynamically to
the photon objects.