Moved to a more declarative option definition.
This just uses standard C-preprocessor stuff so that we (hopefully) don't have
to muck around in C source code to add new options. We should be able to simply
use the options.ggo and options.def files to add new options. As we expand the
options infrastructure we'll probably need a bit more work on this.