Fixed count-based termination detection.
The counts were read from an array reduced over the localities.  The
index for the finished count should have been 1, but 0 was used
accidentally.  In the process, a potentially more serious bug was
discovered.  Active count should never be decreased, but in the case
where a vertex had 0 neighbors, it could have been.  The code is
changed now to never decrease active count, and to always increase
finished count at the end of every visit.  As noted in the code, the
finished count increase could be hoisted up at the cost of code
complexity.

In the process, made the output file work better by changing mode to
"w".

Signed-off-by: Marcin Zalewski <zalewski@indiana.edu>