An SSSP psuedo-fix
SSSP failed with the dyamic action registration using action table because
the action names in libPXGL were not entirely unique.

There was a _put_edge action in edge_list.c and adjacency_list.c, and the
latter registration silently clobbered the previous registration.

With action tables, for now, we have introduced a uniqueness constraint
where action names have to be unique across different files. In the future,
we might consider prepending the file names to action names, but that makes
action tracing kind of ugly.


Former-commit-id: 87ed25a95c754bf250ca94b5956ac780800d213e