Parallelized reading of graph inputs.
HPX does not seem to deal well with one locality throwing out a lot of
messages.  To mitigate this, graph file reading is now parallelized.
There is a possibility of having multiple reading threads per
locality, but it seems that one thread does the best.

Computing of the maximum and minimum edge weights is not implemented
currently (it should be eventually).

Furthermore, there are still performance bottlenecks wherever a single
loop that spawns messages is used.  All these loops should be
rewritten in this style.

Signed-off-by: Marcin Zalewski <zalewski@indiana.edu>

Former-commit-id: 8828da58e302336e256282a58173765e6da7a292