Introduce the action table infrastructure.
Presently, HPX was passing function pointers around as actions. This worked so long
as the "symmetric code" constraint was enforced. This forced us to link HPX as a
static library since the symbols (actions) within a shared library could be loaded
at any address randomly. For our ongoing PWC support, we need to be able to encode
actions within a 64-bit key. The action table infrastructure uses a table with 4K
action entries so that an action can be represented with <16 bits.

- Introduced a new LIBHPX_REGISTER_ACTION function.
- Changed the definition of hpx_register_action to accept an action pointer
- Changed the definition of hpx_run to accept an action pointer
- Introduced hpx_finalize_actions.


Former-commit-id: e6ea7416900ecffb13b2cc033d7ab87161bf3f46