Removed "-static" option from Makefile.am files.
The -static option was used in several makefiles.  It does not seem to
serve a purpose, but if it does, feel free to revert.

The problem that the use of this option causes is that it disables the
following:

make LDFLAGS="-all-static"

This is necessary on Cray with static builds to direct libtool to use
static libraries.  If -static is used in the makefiles as it was, it
overrides -all-static because it appears earlier in the options given
to linker commands.

Signed-off-by: Marcin Zalewski <marcin.zalewski@gmail.com>

Former-commit-id: 79267247997273ec9b12ed83b4537c747c55dbdd