AND LCO fix.
We had an optimization in AND LCOs where we used atomic fetch-and-add +
load to avoid excessive locking. This, sometimes, led to incorrect
synchronization between two threads operating on an AND LCO, where
a set was happening on a deleted LCO.

The fix is to conservatively hold a lock around the operations instead
of using the atomics.


Former-commit-id: 8eaadf14abde1aaaa77b459fc239e3d339442e94