Interface changes without implementation changes.
1) Eliminated the "_class" modifier that we use inconsistently to describe class
interface structures. In general, these are obvious and simply add unneeded
verbosity. The only remaining use of "_class" is the lco_class_t, in which we
already have an lco_t declared as a union.

2) Eliminated some unused constructor parameters for smp() classes. There is no
need for all of the constructors for particular interfaces to have the same
type, since we have to use a switch-based constructor anyway (they are not
virtual or used via function pointers).


Former-commit-id: 2574f1951d734c9af866a0da89e15ec705cbc4ab