Added an option for direct hugetlbfs malloc of the heap.
This option has been only tested on Big Red 2 as for now, and may or
may not work on other systems.

Steps to use:
1. Load any hugepages module on the *compilation* node (usually head
node).
2. Use "--with-hugetlbfs=craype-hugetlbfs" configure option.
3. Do *not* load hugepages on aprun node when using jemalloc.
hugepages can probably be loaded with libc malloc, but that has not
been tested.

To use jemalloc, do not forget:

export CRAYPE_LINK_TYPE=dynamic

Signed-off-by: Marcin Zalewski <marcin.zalewski@gmail.com>

Former-commit-id: a6fe97cdfcb6bfccc42f3aef93991bbffd2500bf