This is a cumulative commit of a new DC and many other improvements.
This work was done on a separate branch, and the commits are
squashed.

The main change is the addition of a new kind of SSSP.  The new kind
of SSSP is DC that maintains separate priority queues per thread.  As
distance messages to vertices come in, the updates are added to the
queue of the thread on which they land.  Such "producers" that add to
the queues, can, with a certain frequency, also empty the queue
becoming "consumers."  There is a threshold that can be set that
attempts to fill queues by running a certain portion of SSSP
asynchronously.  That portion is controlled by a command-line
parameter.  Finally, there is a "flush" light-weight thread that
empties the queues if their size did not increase since it visited a
thread last time.  Currently, the performance of this algorithm is
suboptimal due to imbalance problems.

Furthermore, many changes have been added to make graph generation
faster.

Finally, this commit requires the next commit that introduces the par
for range.

Signed-off-by: Marcin Zalewski <zalewski@indiana.edu>

Former-commit-id: aeb2b03ecc9c6ed2635ab190819f151f67663614