Implement eager buffer overflow.
This uses a basic scheme for eager buffer overflow. When the sender thinks that
it is about to overflow the receiver, it issues an rdma get() operation to read
the receiver's progress. At this point, it switches to buffering mode.

When the rdma completes, we get a local event from photon_probe(), and at that
time we progress the buffer as far as possible, and reissue another get() if
sends remain buffered.


Former-commit-id: 0fe5a7d34a1ac1b0d67cb6ed29f9af9ef2ff67b3