Changed the allocation of heap memory.
Now, heap memory is again allocated through mmap.

We repeatedly try to allocate memory at an address aligned first to 32
bits, and then less by one bit until we reach the chunk size.  The
idea is that we can have full 32-bit blocks where available, and size
down where not available.  There is an assupmtion that our address
space is nto yet trashed, and we should be able to allocate the
addresses we are asking about.  This can break badly if something is
allocated by default in the address range we are asking for, but we
have to see if this is a problem anywhere.

Part of the reason we are using mmap is to enable hugepages mmap
allocation on machines where huge pages are available.  We will
allocate huge pages by using mmap flags set during configuration.

Now we don't need various alignment checks and hacks since we get an
aligned address, so all of these were removed.

Signed-off-by: Marcin Zalewski <zalewski@indiana.edu>

Former-commit-id: 589e96a1ca2262a7feff7b9071df8fd0889af6d0