Fix a longstanding shutdown bug.
In order to shutdown, our cooperative workers check the shutdown flag inside of
their _schedule() loop. When they see a shutdown signal, they transfer back to
their system stacks from inside of _shutdown().

We have to delay this transfer if the lightweight thread calling _schedule() is
holding an LCO lock, because there can be other workers running lightweight
threads that are spinning on this LCO. If we shutdown eagerly they will never be
able to acquire the lock, but they'll also never detect the shutdown signal
because they are in spin-lock loops and we're cooperative.

This patch changes the "fast" scheduler logic (renamed to "in_lco" which is what
"fast" always meant) so that it returns the HPX_ACTION_NULL action when the
scheduler is shut down. This lets the caller release the lock and transfer to
the HPX_ACTION_NULL action, which will exercise the other scheduler path (since
the HPX_ACTION_NULL action isn't in a hurry) and perform the shutdown.


Former-commit-id: 71086ee88cf0661d8fe14a9b579927fa479ad06e