Fix an cyclic heap bug introduced with a new jemalloc.
Jemalloc used to bypass the thread-cache automatically when a specific arena was
specified. They have changed this behavior and require a specific flag to
disable caching. We have always relied on this behavior to manage the cyclic
part of our heap (though not the ret of the global heap, which relies on caching
for performance).


Former-commit-id: efebdfb0c5a1cf80805aeba99baf30ab5d9f4811