Add a comma separator for the foreach macro.
At this time the macro is supposed to preserve the "comma-separated-argument"
semantics of VA_ARGS, essentially, the output is still VA_ARGS, just transformed
slightly. In the future we could consider a version that takes a separator, but
I couldn't figure out how to pass comma as a separator, so here we are.


Former-commit-id: 39f651cd045b0ebb15a68cee84dfa0ed53c02373