Add an hpx_xcall() experimental call macro.
This uses the foreach-functionality to take call arguments directly and prepend
the address of operator to them directly. This technique could reduce bugs, as
the existing call cannot check to see if you've remembered to take an address
of. At the same time, you can't take the address of a temporary so this might
sometimes create a need for a named temporary.


Former-commit-id: d4e064a3c4bdbf4b4fe92a9c457379d592d984f4