Integral constants aren't interpreted as 64 bits, read length as int.
When we are using the old-style actions and passing a pointer and size to the
set of hpx_call actions, we read out the size using va args. We were reading it
as a size_t. When we pass an integral constant like '0' we were getting garbage
out though, because only 32-bits were being written to the stack, so we would
get 32 least-significant bits as 0, but then garbage in the high order
bits. This would cause 'len' to be really large and overflow our allocator
causing the oom.

This fixes clang unit tests (for 3.5.1). It does not appear to be a complete fix
for gcc-4.6.4.


Former-commit-id: e10833165b9b120db1ce8ac741d5dd4e29369a7b