Hack to bypass libffi bug.
All of the libffi function return values seem to be set with a movq
instruction. Unfortunately, this writes 8 bytes. Our actions all return 4 bytes
integers, so we were only allocating 4 bytes on the stack. Then libffi goes and
does this
https://gitlab.crest.iu.edu/extreme/hpx/blob/develop/contrib/libffi/src/x86/unix64.S#L160
where %rdi points to a 4-byte stack value. This clobbers part of the stack.

In gcc-4.6.4, this bit of stack is a frame pointer. Other compilers seem to
accidentally assign a stack slot to the output value such that the 4 extra bytes
don't overwrite live data. I believe this to be a fragile coincidence rather than
anything planned.

This change allocates 8 bytes on the stack but only uses the lower four for the
int. This means that libffi is free to clobber the other dour without doing any
damage.

We should communicate with the libffi developers to see if this is a bug in
their code, or a bug in our understanding of their API.


Former-commit-id: ab7d91cc71fe03492b22ef0e7ba194541b7bb254