Remove HPX_HWLOC_CFLAGS from hpx_extra_cflags.
As such, we shouldn't really need CFLAGS of any of contrib's packages; and
all of them (except HWLOC) were empty anyway. HWLOC set its HWLOC_CFLAGS to
"-fvisibility=hidden" which was used to compile libhpx and resulted in the
default visibility of its symbols to be hidden.

This moves us one step closer towards supporting "--enable-shared."


Former-commit-id: 1be8415bd57784809fe4c60250e0fd6d283dda15