Added a test to see if we're seeing a race in the scheduler.
It seems like we're starting the same parcel on more than one scheduler. They
race to install their own stack, and then go off and execute. This can lead to
all sorts of difficult to debug issues.

Currently, in the test suite, we only see this in the clang 3.5.1 build with
optimizations enabled. This may be affecting larger applications under gcc
though.


Former-commit-id: ec80609075c5f9198a17014b73154c3d59872fa4