A few changes to how we deal with actions:
HPX_ACTION_NULL now is defined to 0, and a separate scheduler_nop action is used
for doing nothing. Newly allocated parcels have their action set to HPX_ACTION_NULL,
instead of this special "nop" parcel.

The action table reserves the first index for HPX_ACTION_NULL so that it is not
given out to other actions.


Former-commit-id: 5dffe2c53d8e490f02f149c3a69ce7aba43c1311