Add a SEGV handler that calls dbg_wait().
This functionality is unreliable in a number of different ways. First of all,
we're replacing MPI's segv handler with our own, which is a potentially
dangerous idea. Secondly, we're calling gethostname and printf inside of a
signal handler, which is also not okay according to the standard. Third, we're
using signal and not sigaction. Fourth, we're running on the stack, so if we get
a SEGV due to a stack overflow bad things will happen.


Former-commit-id: b6a6f3e0c384ec16333a22e24889fd060ed188b0