Recover the lost performance.
Effectively reverted the AND lco fix (8eaadf1), and made AND
use relaxed atomics.

A set is likely to race with a delete only when we are doing too many
sends. If so, we end up with a case where we're trying to set a deleted
LCO, and we safely disregard that as a user-error.


Former-commit-id: b3fd0bdcdc8c8c95d9659920a74fda9b21f7586a