Removed the struct and array capabilities for typed actions.
Currently, libffi performs lazy copies for pass-by-value structures. Lazy in the
sense that it saves a pointer to the structure and doesn't it until the ffi_call
occurs. This presumably has to do with stack allocation being cheaper than
malloc(), and eliminating an extra copy if the same arg structure could be
reused (we need a fresh copy each time according to the semantics of
pass-by-value).

We had expected an eager serialization into libffi's target buffer. Without one,
we can't guarantee that the data is still valid during the lazy call if its done
locally, nor can we send the arg buffer over the network. In the future we could
allow these types to work by taking pack/unpack operators from the user (or
compiler or C++ template expansion) at action registration time. At this point,
we can only support typed actions using native value types.


Former-commit-id: 7e3faa06037fae33a61b412312489d78e9898dbf