Remove transfer instrumentation.
scan-build detects this as generating a garbage value, and it's correct because
the _try_task routine does something different than other transfers. It turns
out this is complicated enough to fix that it's easier to take it out for now. I
left the debug_transfer code available for anyone that needs it.


Former-commit-id: 674b075d60cf35f97f8a300884d208327189a0a1