Export mmap in the system header.
We will prefer this implementation since mmap() tends to be extremely system
dependent. At some point this will encapsulate huge-page functionality as
well. The "align" parameter to system_mmap() is unused at the moment; it's not
clear if it is a better idea to have the client suggest an aligned "addr"
parameter, or use align and do it ourselves.


Former-commit-id: c93ec75d90ac554b26bdef6adac2575d1cebed3c