Stop setting the pthread stack size.
It's become clear that it is difficult to actually say how much pthread stack we
need to run tasks. The system is doing weird thing with them, so main is
actually called quite a distance in.

I'm reverting this optimization because I'm worried it will lead to stack
overflows that can't be traced.


Former-commit-id: 8ec2a3aeb1f18ce300018bc8f6216733dd554bf0