Fix broken pinned-typed actions.
Pinned typed actions are registered with an extra implicit
argument, HPX_POINTER, and libffi expects this parameter to
be present when packing the arguments for a pinned typed
action.

We send over a junk value for the pointer which is rewritten
at the destination with the translated pinned target address.


Former-commit-id: 01ea6846658ac54b6f7ef75368af8f0499bc4710