Remove visibility protection on "here" so that we can unit test bootstrap.
This is not how we should do this. There are a couple of outstanding issues on
visibility in libhpx. We should really be building a "Test" version of libhpx
along with the normal version, where we #define HPX_INTERNAL differently, or use
a linker script to export symbols, so that we can unit test.


Former-commit-id: b9c6bf66582210bb20a26be9c8af4fede8a78d22