Fix initialization of arrays in parcel emulation.
Because I am a masochist and want to pretend that I don't know what the
underlying transport key type or size is, I am using flexible array members to
store them in the parcel emulation layer. They are stored in arrays, but I can't
use normal array indexing because that doesn't account for the key.

I switched the array types to be `char*` so that I wouldn't accidently index
into one, and then wrote a couple of accessors to compute the "right" index
locations for these arrays. This clears up a bug that I thought was all-to-all
related before.


Former-commit-id: 0eb47f454d0626147e5d226105c7f5cdf7616b6f