Add the ability to "retain" parcels and to "nest" parcels a single level deep.
This allows us to avoid copying parcels sent as data during attach. For
applications that depend on attach this can be important. It requires some
complication of the parcel state in order to interoperate with atomics properly.

At this point it might make sense to define state "bits" as flags rather than as
a bitfield, but it's okay for now.


Former-commit-id: a3df904b68df5e2e39a81dd4f9a12f5ec021fc3a