Abstract the eager parcel buffers as a parcel_block_t.
We now allocate parcel buffers as parcel_block_t buffers, with a header that
retains the number of remaining bytes. We set the initial buffer `i` value
according to this header, and that gets transmitted as part of the reload reply.

This will let us reference count buffers so that we can process parcels inline.


Former-commit-id: 4289d19b90030d8e4e3fe587453a1210bf83cbe6