Export the network command details in the network header.
The command type is really part of the network interface and should be exported
there, along with some general commands that are useful inside and outside of
the network code.

I did not export command packing as that it not yet necessary, but it could be
exported in the future.


Former-commit-id: 40c6ad006413423b0a5c730db1f32a8f651c9256