Don't try and set each thread's affinity.
Both aprun and mpirun have mechanisms for controlling that affinity of the
threads that are spawned. Overriding this affinity is genarally not the right
thing do---it both breaks the user's expectations and causes huge startup delays
on Cray (at least our current implementation does).

Removing this code will casue HPX to run poorly if the user launches it
poorly. We need to add debugging code to make sure that our cpu set basically
matches the number of threads that the user has asked us to create.


Former-commit-id: b3b764af74b883c7c561dbb925c44d25b5f76ac7