Switch the photon pwc transport to use the huge-pages munmap.
This deals with the problem we have when the user is electing to use huge pages
that are larger than jemalloc's chunk size, without trying to munmap too much
space for other allocations that came from the normal system_mmap().

We currently don't have cases where we will be using huge-pages sometimes but
not others, but we could in the future.


Former-commit-id: a7252ae20a5d53a3736b10cac08f8f300d88716f