Skip the LCO chan perf test.
Now that we have changed the default number of threads/cores to be determined
by the affinity mask (and not by the maximum number of available cores like
we did), comment out the perf test which needs 2 threads by default. This avoid
a failure when running "make check" interactively without an allocation.


Former-commit-id: c0f63189a5b362c48756cb2dd5ffb77fc3cbd20e