Added an --{enable,disable}-pwc option to configure.
Presently the availability of a transport, MPI (--with-mpi) or
Photon (--enable-photon), determines which network we build. In
the future, when a network supports multiple transports we would
want to selectively configure which network we want to build.

For the Photon/UD ISIR transport, we want to disable photon-pwc as
both require linking against different branches of photon. This
change enables us to --disable-pwc when it is not required.


Former-commit-id: a8e15755f31a9db7c0846453bfc7d31d82302208