Update the LCOs that used to store action ids rather than function pointers.
HPX is designed around an AGAS model, which means that LCOs can possibly be
relocated between address spaces. Due to system issues like randomization and
dynamic-library loading function pointers are not portable across localities,
this is why we have the action registration interface. Any function pointer that
may be passed between localities must be registered as a function and be stored
as an action ID.

1 Updated LCOs that use function pointers to use actions instead.
2 Added an HPX_FUNCTION_DEF macro to register "global" functions.
3 Updated all of the in-tree uses of monoids and predicates to use actions.


Former-commit-id: 8ea46ff9438b97687383c6c6195b5cb96a8bd508