Added whitespace around the comma preceding ##.
This is mentioned in the gcc documentation which states:

Previous versions of CPP implemented the comma-deletion extension much more generally. We have restricted it in this release to minimize the differences from C99. To get the same effect with both this and previous versions of GCC, the token preceding the special ‘##’ must be a comma, and there must be white space between that comma and whatever comes immediately before it:

>  #define eprintf(format, args...) fprintf (stderr, format , ##args)


Former-commit-id: 12a7644b52aad59a508b6c64f70e68f884f9a0a7